+++ src/app/app.module.ts
@@ -16,25 +16,27 @@ import { LoginComponent } from './login/login.component';
 import { ToolbarComponent } from './toolbar/toolbar.component';
 
 import { PageNotFoundComponent }  from './page-not-found/page-not-found.component';
+import { AppRoutingModule } from './app-routing.module';
 
-const appRoutes: Routes = [
-  // delete data field because it is reduntant
-  { path: '', redirectTo: 'login', pathMatch: 'full' },
-  {
-    path: 'login', component: LoginLayoutComponent, data: {title: 'First Component'},
-    children: [
-      {path: '', component: LoginComponent}
-    ]
-  },
-  { path: 'main', component: HomeLayoutComponent,
-    children: [
-      { path: 'first', component: FirstComponent },
-      { path: 'second', component: SecondComponent },
-      // angular route should have a '**' route for 'page not found' -- found Angular.io tutorial
-      { path: '**', component: PageNotFoundComponent}
-    ]
-  }
-];
+
+// const appRoutes: Routes = [
+//   // delete data field because it is reduntant
+//   { path: '', redirectTo: 'login', pathMatch: 'full' },
+//   {
+//     path: 'login', component: LoginLayoutComponent, data: {title: 'First Component'},
+//     children: [
+//       {path: '', component: LoginComponent}
+//     ]
+//   },
+//   { path: 'main', component: HomeLayoutComponent,
+//     children: [
+//       { path: 'first', component: FirstComponent },
+//       { path: 'second', component: SecondComponent },
+//       // angular route should have a '**' route for 'page not found' -- found Angular.io tutorial
+//       { path: '**', component: PageNotFoundComponent}
+//     ]
+//   }
+// ];
 @NgModule({
   declarations: [
     AppComponent,
@@ -45,7 +47,8 @@ const appRoutes: Routes = [
     HomeLayoutComponent,
     LoginComponent,
     ToolbarComponent,
-    PageNotFoundComponent
+    PageNotFoundComponent,
+    AppRoutingModule
   ],
   imports:[
  CommonModule,
@@ -53,10 +56,11 @@ NgtUniversalModule,
 
     BrowserModule,
     BrowserAnimationsModule,
-    RouterModule.forRoot(
-      appRoutes,
-      { useHash: false } // <-- debugging purposes only
-    ),
+
+    // RouterModule.forRoot(
+    //   appRoutes,
+    //   { useHash: false } // <-- debugging purposes only
+    // ),
     CustomMaterialModule
   ],
   providers: [],
